+++ common/datavalues/src/arrays/binary/mod.rs
@@ -12,10 +12,119 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod binary;
 mod builder;
 mod iterator;
 
-pub use binary::*;
 pub use builder::*;
+use common_arrow::arrow::array::*;
+use common_arrow::arrow::bitmap::Bitmap;
+use common_exception::ErrorCode;
+use common_exception::Result;
 pub use iterator::*;
+
+use crate::prelude::*;
+
+#[derive(Debug, Clone)]
+pub struct DFBinaryArray {
+    pub(crate) array: LargeBinaryArray,
+}
+
+impl From<LargeBinaryArray> for DFBinaryArray {
+    fn from(array: LargeBinaryArray) -> Self {
+        Self { array }
+    }
+}
+
+impl DFBinaryArray {
+    pub fn new(array: LargeBinaryArray) -> Self {
+        Self { array }
+    }
+
+    pub fn from_arrow_array(array: &dyn Array) -> Self {
+        Self::new(
+            array
+                .as_any()
+                .downcast_ref::<LargeBinaryArray>()
+                .unwrap()
+                .clone(),
+        )
+    }
+
+    pub fn get_inner(&self) -> &LargeBinaryArray {
+        &self.array
+    }
+
+    pub fn data_type(&self) -> &DataType {
+        &DataType::Binary
+    }
+
+    /// # Safety
+    /// Note this doesn't do any bound checking, for performance reason.
+    pub unsafe fn try_get(&self, index: usize) -> Result<DataValue> {
+        let v = match self.array.is_null(index) {
+            true => None,
+            false => Some(self.array.value_unchecked(index).to_owned()),
+        };
+        Ok(v.into())
+    }
+
+    pub fn len(&self) -> usize {
+        self.array.len()
+    }
+
+    #[inline]
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
+    #[inline]
+    pub fn null_count(&self) -> usize {
+        self.array.null_count()
+    }
+
+    #[inline]
+    pub fn is_null(&self, i: usize) -> bool {
+        self.array.is_null(i)
+    }
+
+    #[inline]
+    /// Get the null count and the buffer of bits representing null values
+    pub fn null_bits(&self) -> (usize, &Option<Bitmap>) {
+        (self.array.null_count(), self.array.validity())
+    }
+
+    /// Take a view of top n elements
+    pub fn limit(&self, num_elements: usize) -> Self {
+        self.slice(0, num_elements)
+    }
+
+    pub fn slice(&self, offset: usize, length: usize) -> Self {
+        let array = self.array.slice(offset, length);
+        Self::new(array)
+    }
+
+    /// Unpack a array to the same physical type.
+    ///
+    /// # Safety
+    ///
+    /// This is unsafe as the data_type may be uncorrect and
+    /// is assumed to be correct in other unsafe code.
+    pub unsafe fn unpack(&self, array: &Series) -> Result<&Self> {
+        let array_trait = &**array;
+        if self.data_type() == array.data_type() {
+            let ca = &*(array_trait as *const dyn SeriesTrait as *const Self);
+            Ok(ca)
+        } else {
+            Err(ErrorCode::IllegalDataType(format!(
+                "cannot unpack array {:?} into matching type {:?}",
+                array,
+                self.data_type()
+            )))
+        }
+    }
+
+    pub fn collect_values(&self) -> Vec<Option<Vec<u8>>> {
+        let e = self.get_inner().iter().map(|c| c.map(|d| d.to_owned()));
+        e.collect()
+    }
+}
