+++ common/datavalues/src/arrays/list/builder.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
 use common_arrow::arrow::array::*;
 use num::Num;
