+++ common/datavalues/src/arrays/ops/take_test.rs
@@ -44,7 +44,7 @@ fn test_take() -> Result<()> {
     assert_eq!(&[Some(true), Some(false)], values.as_slice());
 
     // Test ListArray
-    let mut builder = get_list_builder(&DataType::UInt16, 12, 3);
+    let mut builder = get_list_builder(DataType::UInt16, 12, 3);
     builder.append_series(&Series::new(vec![1_u16, 2, 3]));
     builder.append_series(&Series::new(vec![7_u16, 8, 9]));
     let df_list = &builder.finish();
