+++ common/datavalues/src/arrays/struct/mod.rs
@@ -12,6 +12,105 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod r#struct;
+use common_arrow::arrow::array::*;
+use common_exception::ErrorCode;
+use common_exception::Result;
 
-pub use r#struct::*;
+use crate::prelude::*;
+
+#[derive(Debug, Clone)]
+pub struct DFStructArray {
+    pub(crate) array: StructArray,
+    pub data_type: DataType,
+}
+
+impl From<StructArray> for DFStructArray {
+    fn from(array: StructArray) -> Self {
+        Self::new(array)
+    }
+}
+
+impl DFStructArray {
+    pub fn new(array: StructArray) -> Self {
+        let data_type = array.data_type().into();
+        Self { array, data_type }
+    }
+
+    pub fn from_arrow_array(array: &dyn Array) -> Self {
+        Self::new(
+            array
+                .as_any()
+                .downcast_ref::<StructArray>()
+                .unwrap()
+                .clone(),
+        )
+    }
+
+    pub fn data_type(&self) -> &DataType {
+        &self.data_type
+    }
+
+    pub fn get_inner(&self) -> &StructArray {
+        &self.array
+    }
+
+    /// # Safety
+    /// Note this doesn't do any bound checking, for performance reason.
+    pub unsafe fn try_get(&self, index: usize) -> Result<DataValue> {
+        let nested_array = self.array.values()[index].clone();
+        let series = nested_array.into_series();
+        let scalar_vec = (0..series.len())
+            .map(|i| series.try_get(i))
+            .collect::<Result<Vec<_>>>()?;
+        Ok(DataValue::Struct(scalar_vec))
+    }
+
+    pub fn len(&self) -> usize {
+        self.array.len()
+    }
+
+    #[inline]
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
+    #[inline]
+    pub fn null_count(&self) -> usize {
+        self.array.null_count()
+    }
+
+    #[inline]
+    pub fn is_null(&self, i: usize) -> bool {
+        self.array.is_null(i)
+    }
+
+    /// Take a view of top n elements
+    pub fn limit(&self, num_elements: usize) -> Self {
+        self.slice(0, num_elements)
+    }
+
+    pub fn slice(&self, offset: usize, length: usize) -> Self {
+        let array = self.array.slice(offset, length);
+        Self::new(array)
+    }
+
+    /// Unpack a array to the same physical type.
+    ///
+    /// # Safety
+    ///
+    /// This is unsafe as the data_type may be uncorrect and
+    /// is assumed to be correct in other unsafe code.
+    pub unsafe fn unpack(&self, array: &Series) -> Result<&Self> {
+        let array_trait = &**array;
+        if self.data_type() == array.data_type() {
+            let ca = &*(array_trait as *const dyn SeriesTrait as *const Self);
+            Ok(ca)
+        } else {
+            Err(ErrorCode::IllegalDataType(format!(
+                "cannot unpack array {:?} into matching type {:?}",
+                array,
+                self.data_type()
+            )))
+        }
+    }
+}
