+++ query/src/servers/clickhouse/writers/query_writer.rs
@@ -166,110 +166,79 @@ pub fn to_clickhouse_block(block: DataBlock) -> Result<Block> {
         let field = block.schema().field(column_index);
         let name = field.name();
         let is_nullable = field.is_nullable();
-        result = match is_nullable {
-            true => match column.data_type() {
-                DataType::Int8 => result.column(name, column.i8()?.collect_values()),
-                DataType::Int16 => result.column(name, column.i16()?.collect_values()),
-                DataType::Int32 => result.column(name, column.i32()?.collect_values()),
-                DataType::Int64 => result.column(name, column.i64()?.collect_values()),
-                DataType::UInt8 => result.column(name, column.u8()?.collect_values()),
-                DataType::UInt16 => result.column(name, column.u16()?.collect_values()),
-                DataType::UInt32 => result.column(name, column.u32()?.collect_values()),
-                DataType::UInt64 => result.column(name, column.u64()?.collect_values()),
-                DataType::Float32 => result.column(name, column.f32()?.collect_values()),
-                DataType::Float64 => result.column(name, column.f64()?.collect_values()),
-                DataType::Date32 => result.column(name, column.date32()?.collect_values()),
-                DataType::Date64 => result.column(name, column.date64()?.collect_values()),
-                DataType::Utf8 => result.column(name, column.utf8()?.collect_values()),
-                DataType::Boolean => {
-                    let v: Vec<Option<u8>> = column
-                        .bool()?
-                        .downcast_iter()
-                        .map(|f| f.map(|v| v as u8))
-                        .collect();
+        result =
+            match is_nullable {
+                true => match column.data_type() {
+                    DataType::Int8 => result.column(name, column.i8()?.collect_values()),
+                    DataType::Int16 => result.column(name, column.i16()?.collect_values()),
+                    DataType::Int32 => result.column(name, column.i32()?.collect_values()),
+                    DataType::Int64 => result.column(name, column.i64()?.collect_values()),
+                    DataType::UInt8 => result.column(name, column.u8()?.collect_values()),
+                    DataType::UInt16 => result.column(name, column.u16()?.collect_values()),
+                    DataType::UInt32 => result.column(name, column.u32()?.collect_values()),
+                    DataType::UInt64 => result.column(name, column.u64()?.collect_values()),
+                    DataType::Float32 => result.column(name, column.f32()?.collect_values()),
+                    DataType::Float64 => result.column(name, column.f64()?.collect_values()),
+                    DataType::Utf8 => result.column(name, column.utf8()?.collect_values()),
+                    DataType::Boolean => {
+                        let v: Vec<Option<u8>> = column
+                            .bool()?
+                            .into_iter()
+                            .map(|f| f.map(|v| v as u8))
+                            .collect();
 
-                    result.column(name, v)
+                        result.column(name, v)
+                    }
+                    _ => {
+                        return Err(ErrorCode::BadDataValueType(format!(
+                            "Unsupported column type:{:?}",
+                            column.data_type()
+                        )));
+                    }
+                },
+                false => {
+                    match column.data_type() {
+                        DataType::Int8 => result
+                            .column(name, column.i8()?.get_inner().values().as_slice().to_vec()),
+                        DataType::Int16 => result
+                            .column(name, column.i16()?.get_inner().values().as_slice().to_vec()),
+                        DataType::Int32 => result
+                            .column(name, column.i32()?.get_inner().values().as_slice().to_vec()),
+                        DataType::Int64 => result
+                            .column(name, column.i64()?.get_inner().values().as_slice().to_vec()),
+                        DataType::UInt8 => result
+                            .column(name, column.u8()?.get_inner().values().as_slice().to_vec()),
+                        DataType::UInt16 => result
+                            .column(name, column.u16()?.get_inner().values().as_slice().to_vec()),
+                        DataType::UInt32 => result
+                            .column(name, column.u32()?.get_inner().values().as_slice().to_vec()),
+                        DataType::UInt64 => result
+                            .column(name, column.u64()?.get_inner().values().as_slice().to_vec()),
+                        DataType::Float32 => result
+                            .column(name, column.f32()?.get_inner().values().as_slice().to_vec()),
+                        DataType::Float64 => result
+                            .column(name, column.f64()?.get_inner().values().as_slice().to_vec()),
+                        DataType::Utf8 => {
+                            let vs: Vec<&str> = column.utf8()?.into_no_null_iter().collect();
+                            result.column(name, vs)
+                        }
+                        DataType::Boolean => {
+                            let vs: Vec<u8> = column
+                                .bool()?
+                                .into_no_null_iter()
+                                .map(|c| c as u8)
+                                .collect();
+                            result.column(name, vs)
+                        }
+                        _ => {
+                            return Err(ErrorCode::BadDataValueType(format!(
+                                "Unsupported column type:{:?}",
+                                column.data_type()
+                            )));
+                        }
+                    }
                 }
-                _ => {
-                    return Err(ErrorCode::BadDataValueType(format!(
-                        "Unsupported column type:{:?}",
-                        column.data_type()
-                    )));
-                }
-            },
-            false => match column.data_type() {
-                DataType::Int8 => result.column(
-                    name,
-                    column.i8()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::Int16 => result.column(
-                    name,
-                    column.i16()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::Int32 => result.column(
-                    name,
-                    column.i32()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::Int64 => result.column(
-                    name,
-                    column.i64()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::UInt8 => result.column(
-                    name,
-                    column.u8()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::UInt16 => result.column(
-                    name,
-                    column.u16()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::UInt32 => result.column(
-                    name,
-                    column.u32()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::UInt64 => result.column(
-                    name,
-                    column.u64()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::Float32 => result.column(
-                    name,
-                    column.f32()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::Float64 => result.column(
-                    name,
-                    column.f64()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::Date32 => result.column(
-                    name,
-                    column.date32()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::Date64 => result.column(
-                    name,
-                    column.date64()?.downcast_ref().values().as_slice().to_vec(),
-                ),
-                DataType::Utf8 => {
-                    let vs: Vec<&str> =
-                        column.utf8()?.downcast_iter().map(|c| c.unwrap()).collect();
-                    result.column(name, vs)
-                }
-                DataType::Boolean => {
-                    let vs: Vec<u8> = column
-                        .bool()?
-                        .downcast_iter()
-                        .map(|c| match c {
-                            Some(c) => c as u8,
-                            None => 0,
-                        })
-                        .collect();
-                    result.column(name, vs)
-                }
-                _ => {
-                    return Err(ErrorCode::BadDataValueType(format!(
-                        "Unsupported column type:{:?}",
-                        column.data_type()
-                    )));
-                }
-            },
-        }
+            }
     }
     Ok(result)
 }
